Update task index after delete
